+++ config/proxy.ts
@@ -8,12 +8,12 @@
 export default {
   dev: {
     '/jetlinks': {
-      // target: 'http://192.168.3.157:8844/',
-      // ws: 'ws://192.168.3.157:8844/',
+      target: 'http://192.168.3.146:8844/',
+      ws: 'ws://192.168.3.146:8844/',
       // target: 'http://water.zlkjhb.com:9000/jetlinks',
       // ws: 'http://water.zlkjhb.com:9000/jetlinks',
-      ws: 'ws://demo.jetlinks.cn/jetlinks',
-      target: 'http://demo.jetlinks.cn/jetlinks',
+      // ws: 'ws://demo.jetlinks.cn/jetlinks',
+      // target: 'http://demo.jetlinks.cn/jetlinks',
       changeOrigin: true,
       pathRewrite: { '^/jetlinks': '' },
     },
