+++ src/components/GlobalHeader/AvatarDropdown.tsx
@@ -44,7 +44,7 @@ const AvatarDropdown: React.FC<GlobalHeaderRightProps> = (props) => {
   useEffect(() => {
     service.get().subscribe((resp) => {
       setUser(resp);
-      localStorage.setItem('tenants-admin', resp.tenants[0]?.adminMember);
+      // localStorage.setItem('tenants-admin', resp.tenants[0]?.adminMember);
     })
   }, [currentUser]);
