+++ src/pages/analysis/Tenant.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+interface Props {
+
+}
+const TenantAnalysis: React.FC<Props> = (props) => {
+    return (
+        <div>123</div>
+    )
+}
+export default TenantAnalysis;
